Update Buckets list to use ResourceList instead of IndexList
